% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub_censor.R
\name{censor_worker}
\alias{censor_worker}
\title{mgsub_censor worker}
\usage{
censor_worker(string, pattern, censor, split = any(nchar(censor) > 1),
  seed = NULL, ...)
}
\arguments{
\item{string}{a character vector where replacements are sought}

\item{pattern}{Character string to be matched in the given character vector}

\item{censor}{character to use in censoring - see details}

\item{split}{if a multicharacter censor pattern is provided, should it be 
split to preserve original string length}

\item{seed}{optional parameter to fix sampling of multicharacter censors}

\item{\dots}{arguments to pass to regexpr family}
}
\description{
The hard worker doing everything for mgsub_censor
}
