% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fastReplace}
\alias{fastReplace}
\title{Fast escape replace}
\usage{
fastReplace(string, pattern, replacement, ...)
}
\arguments{
\item{string}{a character vector where replacements are sought}

\item{pattern}{Character string to be matched in the given character vector}

\item{replacement}{Character string equal in length to pattern or of length 
one which are a replacement for matched pattern.}

\item{\dots}{arguments to pass to gsub}
}
\description{
Fast escape function for limited case where only one pattern
provided actually matches anything
}
