% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fallback_helpers.R
\name{filter_overlap_base}
\alias{filter_overlap_base}
\title{Filter overlaps from matches using only base R functionality}
\usage{
filter_overlap_base(x)
}
\arguments{
\item{x}{Matrix of gregexpr results, 4 columns, index of column matched,
start of match, length of match, end of match. Produced exclusively from
a worker function in mgsub}
}
\description{
Helper function using only base R functionality
used to identify which results from gregexpr
overlap other matches and filter out shorter, overlapped results
}
