% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filterOverlap}
\alias{filterOverlap}
\title{Filter overlaps from matches}
\usage{
filterOverlap(x)
}
\arguments{
\item{x}{Matrix of gregexpr results, 4 columns, index of column matched, 
start of match, length of match, end of match. Produced exclusively from 
a worker function in mgsub}
}
\description{
Helper function used to identify which results from gregexpr
overlap other matches and filter out shorter, overlapped results
}
