% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getMatches}
\alias{getMatches}
\title{Get all matches}
\usage{
getMatches(string, pattern, i, ...)
}
\arguments{
\item{string}{a character vector where replacements are sought}

\item{pattern}{Character string to be matched in the given character vector}

\item{i}{an iterator provided by a looping function}

\item{\dots}{arguments to pass to gregexpr}
}
\description{
Helper function to be used in a loop to check each pattern
provided for matches
}
