% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{singleSafeReplace}
\alias{singleSafeReplace}
\title{Single safe replace}
\usage{
singleSafeReplace(x, pattern, string, replacement, ...)
}
\arguments{
\item{pattern}{Character string to be matched in the given character vector}

\item{string}{a character vector where replacements are sought}

\item{replacement}{Character string equal in length to pattern or of length 
one which are a replacement for matched pattern.}

\item{\dots}{arguments to pass to sub}
}
\description{
Fast escape function for limited case where only one pattern
should be matched due to overlaps
}
