% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{worker}
\alias{worker}
\title{mgsub worker}
\usage{
worker(string, pattern, replacement, ...)
}
\arguments{
\item{string}{a character vector where replacements are sought}

\item{pattern}{Character string to be matched in the given character vector}

\item{replacement}{Character string equal in length to pattern or of length 
one which are a replacement for matched pattern.}

\item{\dots}{arguments to pass to regexpr family}
}
\description{
The hard worker doing everything for mgsub
}
