% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub-package.R
\docType{package}
\name{mgsub-package}
\alias{mgsub-package}
\title{mgsub: Safe, Multiple, Simultaneous String Substitution}
\description{
Designed to enable simultaneous substitution in strings in a safe fashion. Safe means it does not rely on placeholders (which can cause errors in same length matches).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bmewing/mgsub}
  \item \url{https://bmewing.github.io/mgsub/}
  \item Report bugs at \url{https://github.com/bmewing/mgsub/issues}
}

}
\author{
\strong{Maintainer}: Mark Ewing \email{b.mark.ewing@gmail.com}

Authors:
\itemize{
  \item Mark Ewing \email{b.mark.ewing@gmail.com}
}

}
\keyword{internal}
